Создать файл типов и поправить Header
